Arregla la sincronización de cardio con minutos decimales
"invalid input syntax for type integer: 28.6": durationMinutes era integer en
Supabase y la app manda lo que se escriba. Un reloj devuelve 28.6 minutos, y
redondear a 29 falsea el gasto calórico que sale de ahí, así que la columna
pasa a double precision como distanceKm y caloriesBurned, que ya lo eran. El
campo de duración lo dice ahora con step="0.1".

Un decimal donde la columna es entera no rompe la pantalla —Dexie lo guarda
igual— sino la sincronización entera de esa tabla, y en silencio hasta que
alguien abre el menú de la cuenta. Las repeticiones tenían el mismo agujero:
targetReps y reps salían de un Number(...) sin más y sus columnas son integer.
Ahí sí el entero es el dato correcto, así que se redondea al leer el
formulario, que es lo único que no se salta pegando un valor.

+++ supabase/schema.sql
@@ -198,7 +198,7 @@ create table if not exists "training_cardio_sessions" (
   "dayId" uuid not null,
   "exerciseId" uuid not null,
   "startedAt" timestamptz not null,
-  "durationMinutes" integer not null,
+  "durationMinutes" double precision not null,
   "distanceKm" double precision,
   "caloriesBurned" double precision,
   "notes" text not null,
